Added tests that the files/directories being deleted by the uninstall target are in fact the correct types (i.e. a file or a directory)

+++ Makefile
@@ -227,9 +227,9 @@ install: all
 	cp src/*.h ${INCDIR}/
 
 uninstall:
-	rm -f ${BINDIR}/luvit
-	rm -rf ${LIBDIR}
-	rm -rf ${INCDIR}
+	test -f ${BINDIR}/luvit && rm -f ${BINDIR}/luvit
+	test -d ${LIBDIR} && rm -rf ${LIBDIR}
+	test -d ${INCDIR} && rm -rf ${INCDIR}
 
 bundle: bundle/luvit
